Renamed _registry to _manager Created Director class to hold all the managers Moved AliasManager into the director Reworked directories for consistency
