+++ abacura/plugins/commands/__init__.py
@@ -1,41 +1,15 @@
 from __future__ import annotations
 
 import inspect
-import re
 import shlex
-from datetime import datetime, timedelta
-from typing import List, Dict, TYPE_CHECKING, Callable, Match
+from typing import List, Dict, TYPE_CHECKING, Callable
 
 from rich.markup import escape
 from serum import inject
 from textual import log
-from abacura.mud import OutputMessage
 
 if TYPE_CHECKING:
     from abacura.mud.session import Session
-
-
-class Ticker:
-    def __init__(self, source: object, callback: Callable, seconds: float, repeats: int = -1, name: str = ''):
-        self.source: object = source
-        self.callback: Callable = callback
-        self.seconds: float = seconds
-        self.repeats: int = repeats
-        self.name: str = name
-        self.last_tick = datetime.utcnow()
-        self.next_tick = self.last_tick + timedelta(seconds=self.seconds)
-
-    def tick(self) -> datetime:
-        now = datetime.utcnow()
-        if self.next_tick <= now and self.repeats != 0:
-            # try to keep ticks aligned , so use the last target (next_tick) as the basis for adding the interval
-            self.next_tick = max(now, self.next_tick + timedelta(seconds=self.seconds))
-            self.last_tick = now
-            self.callback()
-            if self.repeats > 0:
-                self.repeats -= 1
-
-        return self.next_tick
 
 
 class Command:
@@ -141,60 +115,8 @@ class Command:
         return "\n".join(help_text)
 
 
-class Action:
-    def __init__(self, source: object, pattern: str, callback: Callable,
-                 flags: int = 0, name: str = '', color: bool = False, priority: int = 0):
-        self.pattern = pattern
-        self.callback = callback
-        self.flags = flags
-        self.compiled_re = re.compile(pattern, flags)
-        self.name = name
-        self.color = color
-        self.source = source
-        self.priority = priority
-        self.parameters = []
-
-        self.parameters = list(inspect.signature(callback).parameters.values())
-
-        self.parameter_types = [p.annotation for p in self.parameters]
-        non_match_types = [Match, OutputMessage]
-        self.expected_match_groups = len([t for t in self.parameter_types if t not in non_match_types])
-
-        valid_type_annotations = [str, int, float, inspect._empty, Match, OutputMessage]
-        invalid_types = [t for t in self.parameter_types if t not in valid_type_annotations]
-
-        if invalid_types:
-            raise TypeError(f"Invalid action parameter type: {callback}({invalid_types})")
-
-
 @inject
-class TickerRegistry:
-    session: Session
-
-    def __init__(self):
-        self.tickers: List[Ticker] = []
-
-    def register_object(self, obj: object):
-        pass
-
-    def unregister_object(self, obj: object):
-        self.tickers = [t for t in self.tickers if t.source != obj]
-
-    def add(self, ticker: Ticker):
-        self.tickers.append(ticker)
-
-    def remove(self, name: str):
-        self.tickers = [t for t in self.tickers if name == '' or t.name != name]
-
-    def process_tick(self):
-        for ticker in self.tickers:
-            ticker.tick()
-            if ticker.repeats == 0:
-                self.tickers.remove(ticker)
-
-
-@inject
-class CommandRegistry:
+class CommandManager:
     session: Session
 
     def __init__(self):
@@ -257,65 +179,3 @@ class CommandRegistry:
             return True
 
         return True
-
-
-@inject
-class ActionRegistry:
-    session: Session
-
-    def __init__(self):
-        self.actions: List[Action] = []
-
-    def register_object(self, obj: object):
-        for name, member in inspect.getmembers(obj, callable):
-            if hasattr(member, "action_pattern"):
-                act = Action(pattern=getattr(member, "action_pattern"), callback=member, source=obj,
-                             flags=getattr(member, "action_flags"), color=getattr(member, "action_color"))
-                self.add(act)
-
-    def unregister_object(self, obj: object):
-        self.actions = [a for a in self.actions if a.source != obj]
-
-    def add(self, action: Action):
-        self.actions.append(action)
-
-    def remove(self, name: str):
-        self.actions = [a for a in self.actions if name == '' or a.name != name]
-
-    def process_output(self, message: OutputMessage):
-        if type(message.message) is not str:
-            return
-
-        act: Action
-        for act in sorted(self.actions, key=lambda x: x.priority, reverse=True):
-            s = message.message if act.color else message.stripped
-            match = act.compiled_re.search(s)
-
-            if match:
-                self.initiate_callback(act, message, match)
-
-    def initiate_callback(self, action: Action, message: OutputMessage, match: Match):
-        g = list(match.groups())
-
-        # perform type conversions
-        if len(g) < action.expected_match_groups:
-            msg = f"Incorrect # of match groups.  Expected {action.expected_match_groups}, got {g}"
-            self.session.output(f"[bold red] # ERROR: {msg} {repr(action)}")
-
-        args = []
-
-        for arg_type in action.parameter_types:
-            if arg_type == Match:
-                value = match
-            elif arg_type == OutputMessage:
-                value = message
-            elif callable(arg_type) and arg_type.__name__ != '_empty':
-                # fancy type conversion
-                value = arg_type(g.pop(0))
-            else:
-                value = g.pop(0)
-
-            args.append(value)
-
-        # call with the list of args
-        action.callback(*args)
