+++ abacura/plugins/loader.py
@@ -13,8 +13,7 @@ from textual.widgets import TextLog
 from abacura import Config
 from abacura.mud.options.msdp import MSDP
 from abacura.plugins import Plugin
-from abacura.plugins.aliases.manager import AliasManager
-from abacura.plugins.registry import TickerRegistry, CommandRegistry, ActionRegistry
+from abacura.plugins.director import Director
 
 if TYPE_CHECKING:
     from abacura.mud.session import Session
@@ -27,10 +26,7 @@ class PluginLoader(Plugin):
     sessions: dict
     session: Session
     msdp: MSDP
-    action_registry: ActionRegistry
-    command_registry: CommandRegistry
-    ticker_registry: TickerRegistry
-    alias_manager: AliasManager
+    director: Director
     tl: TextLog
 
     def __init__(self):
@@ -73,9 +69,7 @@ class PluginLoader(Plugin):
             # Look for plugins subclasses within the module we just loaded and create a PluginHandler for each
             for name, c in inspect.getmembers(module, inspect.isclass):
                 if c.__module__ == module.__name__ and inspect.isclass(c) and issubclass(c, Plugin):
-                    with Context(session=self.session, msdp=self.msdp, config=self.config,
-                                 alias_manager=self.alias_manager, action_registry=self.action_registry,
-                                 command_registry=self.command_registry, ticker_registry=self.ticker_registry):
+                    with Context(session=self.session, msdp=self.msdp, config=self.config, director=self.director):
                         plugin_instance: Plugin = c()
 
                     plugin_name = plugin_instance.get_name()
